requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Django==5.0
djangorestframework==3.14.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mock-open==1.4.0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pytest==7.4.3
PyYAML==6.0.1
requests==2.31.0
resolvelib==1.0.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
